backend/requirements.txt
# backend/requirements.txt
# Install: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Config
python-dotenv==1.0.1

# Data validation
pydantic==2.7.1

# AI / Vector DB
qdrant-client==1.9.1
sentence-transformers==2.7.0

# LLMsdeactivate
google-generativeai==0.5.4
requests==2.31.0

# PDF processing
pdfplumber==0.11.0

# Voice
openai-whisper==20231117
gTTS==2.5.1

# Language detection
langdetect==1.0.9

# HTTP
httpx==0.27.0